Move DeleteRowButton into from components
